+++ index.d.ts
@@ -16,7 +16,7 @@ declare module 'react-native-confirmation-code-input' {
     codeLength?: number;
     defaultCode?: string;
     compareWithCode?: string;
-    whenCompareIgnoreCase?: boolean;
+    ignoreCaseWhenCompareCode?: boolean;
 
     activeColor?: string;
     cellBorderWidth?: number;