+++ jest.config.js
@@ -10,7 +10,12 @@ module.exports = {
       statements: 100,
     },
   },
-  coveragePathIgnorePatterns: ['/node_modules/', '/.setup-tests.js'],
+  coveragePathIgnorePatterns: [
+    '/node_modules/',
+    '/.setup-tests.js',
+    '(.*)/styles.js',
+    '<rootDir>/src/components/ConfirmationCodeInput/validation.js',
+  ],
   transform: {
     '^.+\\.js$': '<rootDir>/node_modules/react-native/jest/preprocessor.js',
   },