+++ src/types.js
@@ -3,6 +3,7 @@
 export type INDEX = number;
 
 export type InputPositions = 'left' | 'right' | 'center' | 'full-width';
+
 export type VariantNames =
   | 'clear'
   | 'border-box'